Reorganize file directories
