+++ src/features/auth/hooks/useAuth.tsx
@@ -2,7 +2,7 @@ import { useAppDispatch } from 'app/hooks';
 import { onAuthStateChanged } from 'firebase/auth';
 import { useEffect } from 'react';
 import auth from 'firebaseConfig';
-import { setUser } from './authSlice';
+import { setUser } from '../authSlice';
 
 function useAuth() {
   const dispatch = useAppDispatch();
